moved Configuration to YGOPro_Launcher.Config
