+++ YGOPro Launcher.csproj
@@ -50,7 +50,7 @@
     <Compile Include="About_frm.Designer.cs">
       <DependentUpon>About_frm.cs</DependentUpon>
     </Compile>
-    <Compile Include="Configuration.cs" />
+    <Compile Include="Config\Configuration.cs" />
     <Compile Include="Customize\Customize_frm.cs">
       <SubType>Form</SubType>
     </Compile>
